Calculating Regularization Path using LARS, also known as LARS_path, using the Wine dataset with LASSO Regression.
+Problem: We use the wine dataset to plot the regularization path of the problem. We compute the values of alphas(lambdas), the
sparse indices (indices of variables that are non-zero) and the co-efficients that are estimated.
We display all the values of the parameters and compute the sum of co-efficients after applying a transpose operation on it,
so we can find the sum vector or co-efficients, where each element of the vector(each sum) is the sum of all the elements of the coefficients
matrix at the same indices.
We normalize the sum vector to range from 0-1 and then plot the regularization path where X = Coefficients.Transpose and
Y = Normalized Sum of Coefficients.